requirements.txt
aiofiles==22.1.0
aiogoogle==4.3.0
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
blinker==1.5
browser-cookie3==0.16.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.3
decorator==5.1.1
exceptiongroup==1.0.1
Faker==14.0.0
feedparser==6.0.10
frozenlist==1.3.1
google-api-core==2.8.2
google-api-python-client==2.56.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
google-auth==2.14.1
googleapis-common-protos==1.56.4
greenlet==1.1.2
gspread-asyncio==1.6.0
gspread==5.4.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
importlib-metadata==5.0.0
jaraco.classes==3.2.3
jeepney==0.8.0
keyring==23.11.0
lxml==4.9.1
lz4==4.0.2
markdownify==0.11.6
MarkupSafe==2.1.1
more-itertools==9.0.0
multidict==6.0.2
nextcord-ext-ipc==2.2.1
nextcord==2.3.2
numpy==1.23.1
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.2.0
pbkdf2==1.3
Pillow==9.2.0
pip==22.3.1
protobuf==4.21.9
psycopg2==2.9.3
pyaes==1.6.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
pygsheets==2.0.5
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pywin32-ctypes==0.2.0
requests-oauthlib==1.3.1
requests==2.28.1
rsa==4.9
SecretStorage==3.3.3
selenium==4.4.0
setuptools==65.5.1
sgmllib3k==1.0.0
shikimori-api==1.0.1
shutup==0.2.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
tonyg-rfc3339==0.1
trio-websocket==0.9.2
trio==0.22.0
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3-secure-extra==0.1.0
urllib3==1.26.12
validators==0.20.0
vk-api==11.9.9
wheel==0.38.4
wsproto==1.2.0
yarl==1.8.1
zipp==3.10.0
